feat: Add symlink creation for .tmux.conf file in dotfiles.yml
This commit introduces a new task in the Ansible playbook to create a symlink for the .tmux.conf file. The task uses the ansible.builtin.file module to create the symlink from the source file located in the playbook directory to the user's home directory. The owner and group of the symlink are set to the user. The state is set to 'link' and the force option is enabled.