Remove auto_inline_edit function from autoedit.sh
The auto_inline_edit function, which used sgpt CLI tool for inline editing of files, has been removed from the autoedit.sh script. This function was responsible for applying changes to a file based on prompts and tags within the file, but it is no longer present in the script. The corresponding option for "Inline Edit" in the autoedit function's case statement has also been removed.